backend/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
faiss-cpu>=1.7.4
python-dotenv>=1.0.0

# Google API dependencies
google-generativeai>=0.8.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.2

# Image processing
pillow>=10.0.0

# HTTP client
requests>=2.31.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0